Split diagnostics JSON tests
